+++ .github/workflows/ci.yml
@@ -1,15 +1,9 @@
 name: CI
 
 on:
-    push:
-        branches:
-            - '*'
     pull_request:
         branches:
             - '*'
-
-permissions:
-    content: write
 
 jobs:
     build:
@@ -22,16 +16,16 @@ jobs:
             - name: Setup Node.js
               uses: actions/setup-node@v4
               with:
-                  node-version: 20
+                  node-version: './nvmrc'
 
             - name: Install dependencies
-              run: npm ci
+              run: yarn install --frozen-lockfile
 
             - name: Lint
-              run: npm run lint
+              run: yarn run lint
 
             - name: Type check
-              run: npm run type-check
+              run: yarn run type-check
 
             - name: Build
-              run: npm run build
+              run: yarn run build
