+++ .yarnrc.yml
@@ -1,1 +1,2 @@
+# Define how Node packages should be installed, disable yarn Plug N play installation strategy
 nodeLinker: node-modules
