+++ index.html
@@ -4,7 +4,7 @@
         <meta charset="UTF-8" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="icon" href="./Logo.ico" />
-        <title>React Base Template</title>
+        <title>Dashboard</title>
     </head>
     <body>
         <div id="root"></div>
