+++ src/components/Loader/Loader.styles.tsx
@@ -1,0 +1,3 @@
+import { CircularProgress, styled } from '@mui/material';
+
+export const StyledLoader = styled(CircularProgress)(() => ({}));
