--- src/components/Loader/Spinner.styles.tsx
@@ -1,3 +1,0 @@
-import { CircularProgress, styled } from '@mui/material';
-
-export const SpinnerStyled = styled(CircularProgress)(() => ({}));
