+++ src/context/Customer/Customer.context.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+import { Customer } from './Customer.types';
+export const CustomerContext = createContext<{ customers: Customer[] } | null>(
+    null,
+);
