+++ src/context/Customer/index.ts
@@ -1,0 +1,2 @@
+export { CustomerContext } from './Customer.context';
+export type { Customer } from './Customer.types';
