+++ src/context/Product/Product.context.ts
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+
+import { Product } from './Product.types';
+export const ProductContext = createContext<{ products: Product[] } | null>(
+    null,
+);
