+++ src/context/Product/index.ts
@@ -1,0 +1,2 @@
+export { ProductContext } from './Product.context';
+export type { Product } from './Product.types';
