+++ src/context/Transaction/Transaction.context.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+import { Transaction } from './Transaction.types';
+
+export const TransactionContext = createContext<{
+    transactions: Transaction[];
+} | null>(null);
