+++ src/context/User/User.context.ts
@@ -2,4 +2,4 @@ import { createContext } from 'react';
 
 import { User } from './User.types';
 
-export const userContext = createContext<{ user: User } | null>(null);
+export const UserContext = createContext<{ user: User } | null>(null);
