+++ src/context/User/index.ts
@@ -1,0 +1,2 @@
+export { UserContext } from './User.context';
+export type { User } from './User.types';
