--- src/context/customer/customer.context.ts
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-
-import { Customer } from './customer.types';
-export const customerContext = createContext<{ customers: Customer[] } | null>(
-    null,
-);
