--- src/context/product/product.context.ts
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-
-import { Product } from './product.types';
-export const productContext = createContext<{ products: Product[] } | null>(
-    null,
-);
