--- src/context/transaction/transaction.context.ts
@@ -1,7 +1,0 @@
-import { createContext } from 'react';
-
-import { Transaction } from './transaction.types';
-
-export const transactionContext = createContext<{
-    transactions: Transaction[];
-} | null>(null);
