--- src/layout/Main.layout.tsx
@@ -1,9 +1,0 @@
-// TODO : DJ_A2_00 Code for Layout
-
-import { Outlet } from 'react-router-dom';
-
-export const MainLayout = () => (
-    <>
-        <Outlet />
-    </>
-);
