+++ src/layout/Main/Main.layout.tsx
@@ -1,0 +1,11 @@
+// TODO : DJ_A2_00 : Code for Layout
+
+import { Outlet } from 'react-router-dom';
+
+const Main = () => (
+    <>
+        <Outlet />
+    </>
+);
+
+export { Main as MainLayout };
