+++ src/pages/Dashboard/Dashboard.pages.tsx
@@ -1,4 +1,4 @@
-// TODO : DJ_A2_00 Implementation of Dashboard Page
+// TODO : DJ_A2_00 : Implementation of Dashboard Page
 import { Box, Button, Typography } from '@mui/material';
 export const Dashboard = () => (
     <>
