--- src/pages/error/Error.pages.tsx
@@ -1,3 +1,0 @@
-// TODO : DJ_A2_00 Implementation of ErrorPage
-
-export const ErrorPage = () => <>this is ErrorPage</>;
