+++ src/pages/index.ts
@@ -1,3 +1,1 @@
-export { Dashboard } from './dashboard/Dashboard.pages';
-export { NotFoundPage } from './not-found/NotFound.pages';
-export { ErrorPage } from './error/Error.pages';
+export { Dashboard } from './Dashboard';
