--- src/pages/not-found/NotFound.pages.tsx
@@ -1,3 +1,0 @@
-// TODO : DJ_A2_00 Implementation of NotFoundPage (404)
-
-export const NotFoundPage = () => <>this Notfound Page</>;
