+++ src/provider/Customer/Customer.provider.tsx
@@ -1,0 +1,9 @@
+import { CustomerContext } from 'context';
+
+export const CustomerContextProvider = ({
+    children,
+}: {
+    children: React.ReactNode;
+}) => (
+    <CustomerContext.Provider value={null}>{children}</CustomerContext.Provider>
+);
