+++ src/provider/Customer/index.ts
@@ -1,0 +1,1 @@
+export { CustomerContextProvider } from './Customer.provider';
