+++ src/provider/Product/Prodcut.provider.tsx
@@ -1,0 +1,9 @@
+import { ProductContext } from 'context';
+
+export const ProductContextProvider = ({
+    children,
+}: {
+    children: React.ReactNode;
+}) => (
+    <ProductContext.Provider value={null}>{children}</ProductContext.Provider>
+);
