+++ src/provider/Transaction/Transaction.provider.tsx
@@ -1,0 +1,11 @@
+import { TransactionContext } from 'context';
+
+export const TransactionContextProvider = ({
+    children,
+}: {
+    children: React.ReactNode;
+}) => (
+    <TransactionContext.Provider value={null}>
+        {children}
+    </TransactionContext.Provider>
+);
