+++ src/provider/User/User.provider.tsx
@@ -1,0 +1,7 @@
+import { UserContext } from 'context';
+
+export const UserContextProvider = ({
+    children,
+}: {
+    children: React.ReactNode;
+}) => <UserContext.Provider value={null}>{children}</UserContext.Provider>;
