--- src/provider/customer/Customer.provider.tsx
@@ -1,9 +1,0 @@
-import { customerContext } from 'context';
-
-export const CustomerContextProvider = ({
-    children,
-}: {
-    children: React.ReactNode;
-}) => (
-    <customerContext.Provider value={null}>{children}</customerContext.Provider>
-);
