--- src/provider/product/Prodcut.provider.tsx
@@ -1,9 +1,0 @@
-import { productContext } from 'context';
-
-export const ProductContextProvider = ({
-    children,
-}: {
-    children: React.ReactNode;
-}) => (
-    <productContext.Provider value={null}>{children}</productContext.Provider>
-);
