--- src/provider/transaction/Transaction.provider.tsx
@@ -1,11 +1,0 @@
-import { transactionContext } from 'context';
-
-export const TransactionContextProvider = ({
-    children,
-}: {
-    children: React.ReactNode;
-}) => (
-    <transactionContext.Provider value={null}>
-        {children}
-    </transactionContext.Provider>
-);
