--- src/provider/user/User.provider.tsx
@@ -1,7 +1,0 @@
-import { userContext } from 'context';
-
-export const UserContextProvider = ({
-    children,
-}: {
-    children: React.ReactNode;
-}) => <userContext.Provider value={null}>{children}</userContext.Provider>;
